Output authenticity token only if csrf exists
It seems that in tests, the "csrf_meta_tags" output is empty, resulting in a situation where the context is null, resulting in an logging error about setting a null value for an input. Looking at the source, it may be that Rails defaults to turning off forgery protection in test environments, which should effectively mean the absence is okay? Alternatively, we could consider turning on forgery protection in test environments, though that not be ideal.

Curious why this wasn't an issue previously. Maybe because we're compiling Webpack assets for a different environment now? ("test" vs. "development", for example)